mappings/operational_need_translation_test.go
package mappings

import (
	_ "embed"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestOperationalNeedTranslation(t *testing.T) {
	translation, err := OperationalNeedTranslation()

	assert.NoError(t, err)
	assert.NotNil(t, translation)

	tMap, err := translation.ToMap()
	assert.NoError(t, err)
	assert.NotNil(t, tMap)

}

func TestOperationalNeedTranslationVerifyFieldsArePopulated(t *testing.T) {
	translation, err := OperationalNeedTranslation()
	assert.NoError(t, err)
	assert.NotNil(t, translation)

	assert.NoError(t, err)
	assert.NotNil(t, translation)

	assertTranslationFields(t, translation)

}
